Vagrantfile: Increase VM memory requirements to 6GB
Running the os_tempest role on Vagrant VMs often leads to the following
error:

==> ubuntu1604: TASK [galera_client : Add keys (primary keyserver)] ****************************
==> ubuntu1604: ERROR! Unexpected Exception, this is probably a bug: [Errno 12] Cannot allocate memory
==> ubuntu1604: to see the full traceback, use -vvv
==> ubuntu1604: ERROR: InvocationError for command '/bin/bash -c /vagrant/tests/common/test-ansible-functional.sh' (exited with code 250)

As such lets increase the memory requirements to 6GB. Moreover, we
switch to the Bento boxes which come with 65GB of diskspace and they are
all built from a central place. Hopefully, this will bring some
uniformity across the Vagrant VMs. This also removes the
openSUSE Leap 42.2 configuration since it's reached EOL.

Change-Id: Iba357117d19670f245e999ffdeb5c954d76443e8