common-tasks: Do not sync preconfigured /etc/pip.conf file
Zuul places a /etc/pip.conf file on the host as shown below

2018-05-02 11:48:59.945452 | TASK [configure-mirrors : Install /etc/pip.conf configuration]

However, we don't need that since we have our own role to configure pip
which overrides the Zuul one anyway. For the hosts which are not
configured with the pip_install role, this can actually cause problems
so we wipe the file and we let our own role configure it the way we want
to.

Change-Id: I45a900b056db1c8a173f9ed1dae1bd998915e9d4