test-log-collect.sh: Fix log collection for host
The lxc_cmd variable is being set when we collect log information for
containers but it is not unset when we do the same thing on the host. As
a result of which, the host repo and pkg logs were wrong since they were
collected from the last container instead of the host.

Change-Id: If2d0a84b46538e44f383967108ca9f67e91f5054
(cherry picked from commit a6ed5f3dc6db9e3fbf4db1473d0cb1bb261dc9ca)