common-tasks: nodepool: Set openSUSE OBS mirror for OpenStack CI
The OpenStack infra are now mirroring the virtualization and cloud
repositories for openSUSE so we don't need to set the OBS mirror
to the official download.opensuse.org anymore. This also ensures
that the task for checking the presence of /etc/ci/mirror_info.sh
is delegated to the host since that file does not exist in the
containers.

Change-Id: Ia530a5901e3e8ec759ba6c79345cbd9a8b7d2b46
(cherry picked from commit c3cced9a1c367d7cbb1e736f96b338a27c21117c)