Stop linters test from being skipped for docs builds
As the linters job currently inherits from the functional
job, it appears that the 'irrelevant_files' parameter is
merged, causing all jobs to skip when the patch is docs-only.

This splits out the right bits into a base job, and then both
the linters and functional job inherit from there.

Change-Id: If73e152b64ef74d868a143ff24256bdf71adc162