Implement clouds.yaml on test hosts
In order to allow us to make use of the Ansible shade
modules to interact with the running OpenStack environment
in tests, we execute the openstack_openrc role against
the test host.

The include_vars for the follow-on play is moved to make
it easier to find, and the initial play's 'pre_tasks' are
renamed to 'tasks' as there is no 'roles' implementation
in the play.

Change-Id: Iec4f266324d4e5f81d21163f68f8d5c32c34f86d