Vagrantfile: Add openSUSE Leap 15.0 support
openSUSE Leap 15.0 has been released on the 25th of May 2018 so we add
support for it in the Vagrantfile. Moreover, we move away from the Bento
boxes because they are not being updated very often so we spend a lot of
time updating the base OS during tests. Moreover, the current releases
(due to being old) cause troubles like missing systemd-networkd and
vagrant user on openSUSE, old udev on CentOS etc.

Change-Id: I64696e1bf36de1b7e7a6ce206bd61539ae32b1f9