Add galera_address to list of variables
Previously, we had a <service>_galera_address for every project,
however, it looks like the patches to move the creation of the
database inside the roles have worked great, but their testing
patches inside this repo. have dropped the usage of the address
which means that the service is configured with 127.0.0.1 (the
fallback if none is provided).

This patch fixes this by adding that variable so that everything
can go through smoothly.  Therefore, the removal of future
<service>_galera_address variables will not affect things.

Change-Id: Ib180fc3a7cbae28b413094b444f7d528b584c9ed