get-ansible-role-requiremnets: Use refspec to try Gerrit changes
Without this patch, deployers have to mimic zuul structure
to locally test Gerrit patches in-depth, with their specific
references, before they are merged into the main tree.

This is a problem for people wanting to not change the location
of the openstack-ansible folder.

This patches solves the problem by using the refspec
parameter of the ansible git module: A user can then
edit its own ansible-role-requirements for testing
patches.

Change-Id: I6c0fa078da3f32b542d3a3492a0fb79c822a5f8d