test-distro_install-vars: Add new var file for distro installations
When we test distro installations we should install all components from
distro packages in order to ensure that we are not mixing PIP and distro
packages if more than one service is deployed on the same host or
container. As such, we add a new file to override the default 'source'
installation option which can be used by distro_install jobs and tox
environments to test that scenario.

Change-Id: I15d4243e89f8800e42723b0fdf2c5bf59e2f7608
Implements: blueprint openstack-distribution-packages