Use Tempest master when testing roles in master
If we're testing a stable branch, it is best to use the stable
released Tempest version.  However, when working in master, it
is possible that an OpenStack service change which has a
corresponding Tempest change would lead to things failing because
we're deploying master of the service but a tagged version of
Tempest which means the tests will fail.

In order to ensure stability of stable branches, we'll keep using
the tagged release but we'll use master Tempest for testing the
roles.

Change-Id: I347afd358850149b359ae64871a858701755d896