Switch upgrade jobs to be non-voting
At the moment, the jobs are all failing because of a change that
was done in openstack-ansible-plugins to support Ansible 2.5.

The plan is to use the integrated repo to do all of the CI including
upgrading evenutally so we will be disabling this job for now until
we can hit that point (this cycle, hopefully).

Change-Id: If0bf32b10237679248e6b90f9240a3e3ae3967ad