Add shade & python-keystoneclient to the tox venv
In order to reduce the packages required to pip install on to the hosts,
we install the extra library into the ansible-runtime venv so that tasks
interacting with keystone using our keystone module can do so by using
localhost delegation. Using this strategy, the library does not need to
be installed on the target host.

This has been done in the integrated repo in [1] and [2].

[1] https://review.openstack.org/557351
[2] https://review.openstack.org/568141

Change-Id: I672ae0f4adbf54a7d71e06e6715d562069fda877