Add mq_setup common task file
This patch adds the common task file for the mq setup tasks
which will be replicated by the proposal bot between the roles.

It also updates the script that's useful for manual sync to
accommodate the ability to sync the task file. Some documentation
is also added.

This allows sharing of role files from a single location.
It will be used for sharing common tasks of many roles, like
the mq, the dabatase, or service setup.

Without this patch, roles implementing the same kind of files
would eventually drift and be different. This is a problem as
this increases the maintenance burden.

Relates-To: https://review.openstack.org/580696
Co-Authored-By: Jesse Pretorius <jesse.pretorius@rackspace.co.uk>
Change-Id: Ib564450eb1f3d63bb5fdd3763046dd57df519d28