Ensure clouds.yaml is placed correctly
When placing the clouds.yaml file, it needs to be accessible to
the user running ansible. This user is not necessarily root and
for openstack-infra is a non-root user.

To ensure that the file is placed in the right directory and with
the right permissions, we provide the right variable value to
ansible when it runs the role.

Change-Id: I748b509d24b0060908ac75322dd20062a1515169