Remove redundant apt cache update task
There's no reason for this task to be in its own file.
It's not used anywhere but this play. The task also is
not necessary, given that every role does its own updates
when it needs to and the test execution does a cache
update right at the start in order to install the dependent
packages.

All in all, this is a waste of time, so let's just purge it.

Change-Id: I9cc872189b87eb65c4ee7623d4141be41d207f7a