Switch connection/become settings for cinder
Some services run on localhost. For that to work,
we need to sudo. Ideally it should rather be broken
out into its own playbook and use 'connection: local'
but to do that we'd have to know which service it is
and every role's inventory is different. So instead,
we use this rather horrible hack.

Change-Id: I8ce839a61eae5386edfa8171ffe95d22e5ff2c99