Use local connection and become for nova_compute
The nova_compute group contains localhost, but due to
the name being different, it was missed in [1].

When targeting localhost with 'remote_user: root' and
using 'become: yes' we lose the environment variables
for the user running the playbook (eg: USER, HOME).

However, if we use 'connection: local' and 'become: yes'
together, it works properly. We apply this change to
ensure this play also has the correct access to change
things on the host.

[1] https://review.openstack.org/582359
Change-Id: Ia87e02c3b965be6f384d2f64911f0af2e4c2d598