Set lxc_* vars in one place
Currently there are differing settings for the lxc_* variables
between the test-prepare-host.yml playbook and the test-vars.yml
vars_file. Given that the vars file is used by the same play,
we remove the additional settings given to the roles directly
so that the settings are in one place only.

Change-Id: I1c5813a63c5f4056bb1a2ca306b2bfac2c3d2563