Add missing services and collect journal logs
This change adds several services to the collection list which are part
of our deliverables yet we've no logging for. This change also
implements journal log collection when a given service is found which
should help folks better debug issues with various services, especially
as we transition to all journal logging.

Change-Id: Ie6f06eb5b0488005263aed397fab7dedc5a287a0
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>