Implement clouds.yaml for root user
The use of clouds.yaml from the zuul user does not
work for swift due to the mess of delegation and
become. In this patch we ensure that it is also
there for the root user so that the role test can
pass.

This issue does not apply to the integrated repo.

Change-Id: Ia152f9c24e4965f89d9477f4e343428bb04b3c59