Make role test-requirements.txt be additive
Some roles need to make use of extra requirements, but
only for tests. For example, the octavia role needs to
use the octavia client for some test tasks.

Instead of duplicating the whole test requirements file
in that repo, we change the mechanism to make it additive
so that the repo test-requirements are added into the
test runtime venv.

Change-Id: If670a32e79bfb4623fa427925cdc992219742d58