tests-ansible-env-prep: Fix plugin, tests and ops testing on Vagrant
Our Vagrantfile is rsync'ing the role to /vagrant. This breaks the
role test since the $WORKING_DIR no longer matches the name of the
OSA project. As such, when we are using vagrant to test the plugins
tests and ops repositories we end up using the HEAD of the repositories
instead of our local copy. We fix this by introducing a new
OSA_PROJECT_NAME and then we test against this to determine the actual
OSA project we are testing

Change-Id: Ic8b99b4f2d241144e1200ddf7c33742e9bc91c5b