Do not use master branch for role tests
OpenStack services shas are overriden for role tests. They
are referring to a variable, test_branch, to determine which
branch to run. By default it uses TESTING_BRANCH environment
variable as the source of truth for its value.
However, TESTING_BRANCH is unset in tox.ini. The resulting
test_branch's value is master.

This is a problem, as stable/rocky role tests should run
stable/rocky version of openstack services.

This fixes it by making sure the fallback of the env lookup
is stable/rocky, not master.

Change-Id: Ifedfbb803fe3e3ccce8c13389a2f4dd23757bd6e