test-log-collect: Improve rsync performance for log collection
When collecting logs we do not care about file deltas so we can use
the --whole-file option to not perform any kind of diff between the
source and the destination. The --inplace option can also improve
performance by copying files directly on the destination instead of
creating intermediate ones first.

Change-Id: I044cd2d56b9b907e5948be1d3bd11ce4a876b1a2