Clone the correct plugins repo for role tests
Currently the role tests always clone the plugins repo without
specifying a branch. The tests also always use a git clone, even
though in zuul the repo is already cloned.

This patch ensures that the branch is specified when doing the
clone, and also will just symlink to the existing zuul clone if
it is present.

Change-Id: Iaf5ee20e534a8279b5b563c820518bb176f0f4bb