zuul: move non-voting jobs to experimental
The non-voting jobs for Gentoo are new and they are not stable so
we need to leave them as experimental until we have some proven
gate history and maintainership.

The SUSE jobs have been breaking often, however, it seems that they
are still in use, so we're going to move them to experimental
now so that those who would like to try to maintain it can work
on it.

Change-Id: I1e586c68fb7167480a68a421f3211470e92a7003