Fix bashate xargs run
Bashate does not fail for "E004 File did not end with a newline"
properly unless the file is the last in the list when multiple
check files are provided (as we do with xargs). Since the file order
is random, if only 1 or 2 files are failing, this can cause very
random linter failures.

The solution is to use 'xargs -n1' to pass only 1 file to each
bashate process.

See also https://review.openstack.org/#/c/614542/

Change-Id: I2186795510c2327571e7631b45de8a112f4d4445