zuul: create openstack-ansible-linters project-template
As we try to transition to integrated tests, we need to drop the
role jobs as some seem to be broken.  However, we don't have a way
to do this without dropping linting jobs.

This patch allows us to replace openstack-ansible-role-jobs by
openstack-ansible-linters to keep linters and add new jobs as
we need/want.

Change-Id: I669b526fe6e02424801e003f8a6bde89a4245026