Pin nova_git_install_branch to before placement removal
Currently it seems that the stable/stein legacy role tests are using
the master branch for all tests, rather than stable/stein. While I hunt
down the root cause of this bug, we can back port this pin to unblock
https://review.opendev.org/658285 and all the patches that depend
on it.

For role tests to work, we need to pin nova to before the placement
removal. Once we're able to re-implement the role tests or remove
them we can remove this pin.

Change-Id: I58788af6f02a4b339a270dff69a86ce7cdea41d3
(cherry picked from commit f7ce3ed094dafce9e70fbea4c87888af098d0804)