Ensure that TESTING_BRANCH is exported
In order for Ansible to consume the env var, it needs to
be exported. Without the export, TESTING_BRANCH is not
set and test-vars.yml results in defaulting to 'master'
which is not appropriate for stable branches.

This patch is s stop-gap to unblock testing for stable/stein.

Another patch will follow which will force any tests using
this mechanism to fail if the env var is not set.

Change-Id: I1ff58205c2c0346dfe1744fdf3921b9db03c2774