Use copy-links when collecting logs
--safe-links was used, which breaks openstack service collection on
the metal /etc paths, because they are symlinked into the venv like:
/etc/neutron -> ../openstack/venvs/neutron-19.1.0.dev12/etc/neutron

Instead we can use --copy-links which will resolve the symlink and
copy the data.

Note: This is a fix copied from I40905d79efa1bef9ae6a7678515fd07fde1bdd00
in the integrated repo. Since the scripts used are virtually identical
the fix must be applied here too.

Change-Id: I2cb193322cacd135214f5c3c5d0736842c458531