Add _oslodb_login_host to db_setup tasks.
The ansible mysql modules defaults to localhost, which means that the
db_setup tasks fail if the mysql processes are bound to the mgmt
interface.

This change adds an extra variable _oslodb_login_host for the db_setup
tasks which is defaulted to _oslodb_setup_host if a specific value is
not defined.

Change-Id: I531109213953f78263be68968f8e9d2dc2fd382a