Add missing test branch for nova-lxd
This has been missing previously and the master branch of nova-lxd
will have been used mistakenly in tests up to now. As nova-lxd is now
deprecated and the master branch empty, tests on OSA stable branches
fail.

This patch ensures that an appropriate stable branch version of nova-lxd
is used.

Change-Id: Ieaedbe6990bf709c1fa89c4d3f4a8b5ae7080851