Set value for tempest_use_tempestconf in test-vars
This variable is used to conditionally include a task file which runs
the python_venv_build role. This means that the handler in the venv
build role is executed conditionally on tempest_use_tempestconf but in
the context of the outermost playbook, where this var is not in scope.

THis patch ensures that tempest_use_tempestconf is defined in the scope
of the top level test playbook.

Change-Id: I8cd5e5626ecc3cfa40315e2bdf62945b7bb1582b