Pin version of tox used for Centos-7 tests
Tox has virtalenv as a dependancy and will install it using pip during
tox installation if a suitable version is not present.

This patch pins the version of tox to one which is compatible with the
version of virtualenv in the centos-7 python-virtualenv distro package.

Without this patch, tests which include the openstack_hosts role
will install the python-virtualenv distro package on top of the pip
intalled virtualenv from the tox install, and this results in a broken
package and the test fails.

Change-Id: I0c07bc46f5b9a519232620f2d79d6a894f317921