Update sync/doc/requirements.yml for python3 transition
Several patches have been merged directly to OSA repos updating the
docs requirements in-repo, when the patch should have been applied
here and proposal-bot used to update all of the other repos.

This patch updates the value in openstack-ansible-tests which will
ensure that the changes commited within other repos are not reverted
on future runs of the proposal bot, and also to update any repos
where the old values remain.

Change-Id: I9e1842855519a39b928423f01bdea008c4af2bba