Increase timout for linters job to 30mins
This job is given a longer time to run for two reasons, longer is
spent in the job setup becasue of the increased number of repos
in required_projects and more time is required to push the job
logs into swift storage than the previous method.

Jobs are currently failing because these startup/log collect phases
are taking additional time rather than there being a lint issue, so
this patch extends the job time to account for that.

Change-Id: Iab1593cf46bcf80022bb353ed0399baadf49b165