Use pip and virtualenv from host packages rather than get-pip.py
For focal, using pip to install python tools to the host results
in a incompatiblity and subsequent job failure when openstack_hosts
later installs the distro package for virtualenv.

The python_venv_build role then fails to be able to create a virtualenv.

This patch changes the host setup in the tests repo to be the same as that
in the integrated repo.

Change-Id: Ief99ea7e969f67130be68ac437b4b16be80cb088