Use setuptools constraint from global-requirements-pins rather than u-c
With the new pip resolver, if a package is constrainted to two
different versions simultaneously then the package cannot be installed.

setuptools has a constraint in both OSA global-requirement-pins and
openstack upper-constraints, so the installation will fail with the
new resolver.

This patch removes any references to setuptools from openstack u-c
and allows the constraint from the OSA repo to be used.

Change-Id: Ib4c3592184973ba9908a1ba17d935a96448969a1