Run notify setup when setup_host differs
There's a usecase, where all datasets might be the same, except
setup_hosts. For example, when we use isolated rabbitmq cluster for RPC
but shared one for notifications. And we have same set of users
and vhosts there. In current state setup won't be performed, since
vhosts and users are the same.

Change-Id: Ib5c44327d5f1edc25bace79fdd057adabb0b4c14