Workaround linting issue
ansible-lint<5.0.0 will try to install latest ansible as it's part of the
requirements. This brings in an issue, where lint will be broken
because of too modern ansible. We workaround this by specifying version
that won't install ansible-core and thus break linting

We also comment out tempest_git_repo as we want it to be installed
from constraints instead of the master version

Change-Id: I8f24804caa6c242a1ed6cc9adb70de08817c8f5e