Restrict pyOpenSSL to less then 20.0.0
pyOpenSSL 20.0.0 requires newer version of cryptography
that is not going to be installed or conflicts with distro-provided
python3-openssl. To workaround that we define requirement more
preciesly.

We also disable installation of tests repo as a python package
when running tox jobs.

CentOS 7 job is disabled to resolve circular dependency.

Change-Id: I5ac82f70f33a1df1cea527e5138fcba398d00857