Respect depends-on for collection installation
This aims to ensure that collections which are managed by zuul will get
installed and used during testing. This is especially important since
our deliverables are being installed as collections.
As a base we take code from the integrated repo and simplify it a bit
by removing user defined collections.

Change-Id: I92b6b3edd5d781b3ce9292b52773b0a854f910c7